requirements.txt
pandas
numpy
pm4py
openai
python-dotenv
plotly
streamlit
tabulate
matplotlib
seaborn


